+++ src/components/Charts.vue
@@ -23,86 +23,31 @@
     </el-select>
     <el-button type="primary">Genrate Graph</el-button>
     <el-empty description="no data selected"></el-empty>
-    <apexchart
-      type="line"
-      height="350"
-      width="500"
-      :options="chartOptions"
-      :series="series"
-    ></apexchart>
-
 
     <p v-once>{{getX()}}</p> 
 
-    {{scar}}
+ {{scar}}
    
-{{sortDataG()}}
-
-
+<p v-once> {{sortDataG()}}</p>
+<graph/>
 
 </template>
 <script>
-import { ref } from "vue";
+import { ref, reactive } from "vue";
 import { firebaseFireStore } from "@/firebase/database";
+import Graph from "./Graph.vue"
 export default {
+  components: { Graph },
  name: "App",
   setup(){
-       const chartOptions = ref({
-      chart: {
-        height: 350,
-        type: "line",
-        zoom: {
-          enabled: false,
-        },
-      },
-      dataLabels: {
-        enabled: false,
-      },
-      stroke: {
-        curve: "straight",
-      },
-      title: {
-        text: "Product Trends by Month",
-        align: "left",
-      },
-      grid: {
-        row: {
-          colors: ["#f3f3f3", "transparent"], // takes an array which will be repeated on columns
-          opacity: 0.5,
-        },
-      },
-      xaxis: {
-        categories: ["Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep"],
-        title: {
-          text: "Month",
-        },
-      },
-      yaxis: {
-        title: {
-          text: "Desktops",
-        },
-      },
-      legend: {
-        position: "top",
-        horizontalAlign: "right",
-        floating: true,
-        offsetY: -25,
-        offsetX: -5,
-      },
-    });
 
-    const series = ref([
-      {
-        name: "Desktops",
-        data: [10, 41, 35, 51, 49, 62, 69, 91, 148],
-      },
-      {
-        name: "Laptops",
-        data: [20, 61, 75, 71, 69, 72, 89, 111, 168],
-      },
-    ]);
+    const graphData = reactive({      
+      data: [0,0,0,0],
+    })
 
-    
+
+
+
 
         const dataG = ref([
       {
@@ -128,7 +73,7 @@ export default {
 
     ]);
 
-   
+
   function getX() {
 
    firebaseFireStore
@@ -164,7 +109,7 @@ export default {
   }
 
 
-const scar = []
+const scar = reactive([0])
 
     function sortDataG(){
          for (let index in dataG.value){
@@ -175,7 +120,7 @@ const scar = []
 
 
 
-    return { chartOptions, series,dataG,getX, sortDataG, scar};
+    return {dataG,getX, sortDataG, scar, graphData};
   },
 
   }
